Remove two-way binding from Edit Collection form
Previously there was a race condition where the user begins editing the
name (new value), but the collection is refreshed asynchronously (from
the page load) then the newly written (but not saved) title is lost.

The initial value on the text input based on the current title when the
'Edit' button is clicked.