Upgrade react-router
This was a painful one. It took a while to work out why it wasn't
working locally.
Turns out, the react-router context (for the new react context api) is
made per build - I guess there's some kind of uuid that is included in
the build output? - and as a result of that they introduced this change
https://github.com/ReactTraining/react-router/issues/6361 which stumpted
me.

In our dev environment we use the DynamicCdnWebpackPlugin. This
automatically loads npm modules from the CDN (and by doing so includes a
UMD package) however within our app webpack includes the project through
ESM. Causing the mismatch. It took me way too long for this penny to
drop.

[#169726013]