Add block on ./ship for outdated modules
We agreed in retro that it wouldn't be end of the world for a ship to be
stopped if there are outdated packages.

It's easy to get around, in the case that you don't have time - but you
should try to keep it up to date!

It'll timeout after 5 seconds, so if you walk away from a ship it'll
just push