Use Hooks for VideoDetailsView
- Had a bunch of fun around continuous useEffect. We updated the
location, which caused a rerender/rerun of the useEffect, which updated the
location, which caused a rerender/rerun of the useEffect, which updated the
location, which caused a rerender/rerun of the useEffect. Which we got
bored of.... When using hooks it's important to check the dependencies
carefully, and ensure that any side effects have been accounted for.

[#169702223]
Signed-off-by: Antony O'Neill <antony@boclips.com>