WIP - Upgrading antd to 4.0.0
Following on from the Icon upgrade work done by Alex, this WIP contains
the majority of the upgrade path. Unfortunately there are still ~30
failing tests, and several breaking design changes.

I've compiled a list of risks, and hurdles on a slab which has some
screenshots that show the design changes as a result.

https://boclips.slab.com/posts/upgrading-antd-3-%E2%86%92-4-sjprc881

[#171673201]