Fix save animation when adding video to collection
This was a nasty bug related to how we use getDerivedStateFromProps,
where sometimes the isSaving state was overriden incorrectly.
See more at:
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html
I haven't figured out a way how to not use the derived state logic, but
this fix should take care of only resetting the isSaving state when the
update actually happened.

[#172027841]