Fix flicker when changing subjects
The underlying for this behaviour is that the shared component to render
any kind of collection tile owns the data and does not have a notion of
old and new data.

To really fix this, we need to rewrite the component. Alex and decided
to instead fix the problem by changing the state of the store before the
request is made. This means that there are no collections in the store
while we are waiting for the response.

A flicker remains as it goes from no collections, to the correct ones.
We could fix this with a skeleton...

[#172025188]