Remove `share=true` logic
Now we always ask for a share code if a use is anonymous. If they
previously (in the same session) provided a shareCode we revalidate that
without asking for the code again. So when they visit a shared
collection, they do not have to provide the code multiple times when
they are browsing it, opening videos etc.

[#173916773]